feat: move entity page components up a level
