+++ apps/web/design-system/slide-up.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import cx from 'classnames';
 import { AnimatePresence, motion } from 'framer-motion';
