+++ apps/web/design-system/spinner.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { motion } from 'framer-motion';
 
 import * as React from 'react';
